Product/requirement.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
requests
aiokafka
